doc and electric field

+++ slothpy/_general_utilities/_math_expresions.py
@@ -355,8 +355,8 @@ def _validate_and_compute_partition_product_and_magnetisation_sum(partition_arra
     sorted_partition_array = take_along_axis(partition_array, sorted_indices, axis=0)
     sorted_momenta_array = take_along_axis(momenta_array, sorted_indices, axis=0)
 
-    partition_product = ones(partition_array.shape[1])
-    magnetisation_sum = ones(momenta_array.shape[1])
+    partition_product = ones(partition_array.shape[1], dtype=sorted_momenta_array.dtype)
+    magnetisation_sum = ones(momenta_array.shape[1], dtype=sorted_momenta_array.dtype)
 
     _compute_products_and_sums(sorted_partition_array, sorted_momenta_array, partition_product, magnetisation_sum)
